# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MetroFlow AI offline traffic engineering advisory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TrafficUploadInput {
  filename: string;
  /** @minimum 0 */
  rows: number;
  /** UTF-8 CSV content. Omitted only for metadata-only validation. */
  content?: string;
  hasRequiredColumns?: boolean;
}

export type UploadValidationStatus = typeof UploadValidationStatus[keyof typeof UploadValidationStatus];


export const UploadValidationStatus = {
  valid: 'valid',
  warning: 'warning',
  error: 'error',
} as const;

export interface TrafficPreviewRow {
  timestamp: string;
  intersectionId: string;
  approach: string;
  volume: number;
}

export interface ProjectSummary {
  name: string;
  corridor: string;
  datasetLabel: string;
  updatedAt: string;
}

export interface DemoKpis {
  intersections: number;
  peakWindow: string;
  worstVc: number;
  worstIntersection: string;
  averageDelay: number;
  recommendationCount: number;
}

export type HourlyVolumePeak = typeof HourlyVolumePeak[keyof typeof HourlyVolumePeak];


export const HourlyVolumePeak = {
  am: 'am',
  midday: 'midday',
  pm: 'pm',
  off: 'off',
} as const;

export interface HourlyVolume {
  hour: string;
  volume: number;
  peak: HourlyVolumePeak;
}

export type IntersectionAnalysisStatus = typeof IntersectionAnalysisStatus[keyof typeof IntersectionAnalysisStatus];


export const IntersectionAnalysisStatus = {
  critical: 'critical',
  watch: 'watch',
  stable: 'stable',
} as const;

export interface IntersectionAnalysis {
  id: string;
  name: string;
  order: number;
  vc: number;
  los: string;
  delay: number;
  queue: number;
  lanes: number;
  demand: number;
  capacity: number;
  approach: string;
  lat: number;
  lng: number;
  status: IntersectionAnalysisStatus;
}

export type PhaseTimingStatus = typeof PhaseTimingStatus[keyof typeof PhaseTimingStatus];


export const PhaseTimingStatus = {
  adjusted: 'adjusted',
  locked: 'locked',
  held: 'held',
} as const;

export interface PhaseTiming {
  phase: string;
  movement: string;
  currentGreen: number;
  recommendedGreen: number;
  minPed: number;
  yellow: number;
  status: PhaseTimingStatus;
}

export type ActivityTone = typeof ActivityTone[keyof typeof ActivityTone];


export const ActivityTone = {
  success: 'success',
  warning: 'warning',
  info: 'info',
} as const;

export interface Activity {
  title: string;
  detail: string;
  time: string;
  tone: ActivityTone;
}

export interface MetroflowDemo {
  project: ProjectSummary;
  kpis: DemoKpis;
  hourlyVolume: HourlyVolume[];
  intersections: IntersectionAnalysis[];
  phases: PhaseTiming[];
  assumptions: string[];
  activities: Activity[];
}

export interface UploadValidation {
  status: UploadValidationStatus;
  rowsProcessed: number;
  warnings: string[];
  errors: string[];
  preview: TrafficPreviewRow[];
  analysis?: MetroflowDemo;
}

export interface SimulationInput {
  /**
     * @minimum 0.7
     * @maximum 1.3
     */
  demandMultiplier: number;
  baselineDelay?: number;
  baselineQueue?: number;
  baselineThroughput?: number;
  recommendedDelay?: number;
  recommendedQueue?: number;
  recommendedThroughput?: number;
}

export interface SimulationMetrics {
  delay: number;
  queue: number;
  los: string;
  throughput: number;
  reduction: number;
}

export interface SimulationResult {
  demandMultiplier: number;
  baseline: SimulationMetrics;
  recommended: SimulationMetrics;
  limitations: string;
}

export interface RationaleInput {
  intersectionId: string;
  peakWindow: string;
  vcRatio: number;
  cycleBefore: number;
  cycleAfter: number;
  phaseBefore: number;
  phaseAfter: number;
  delayBefore: number;
  delayAfter: number;
  safetyStatus: string;
  demand?: number;
  capacity?: number;
  queue?: number;
  currentSignal?: string;
  recommendedSignal?: string;
  simulationBefore?: SimulationMetrics;
  simulationAfter?: SimulationMetrics;
  assumptions?: string[];
}

export type RationaleResultSource = typeof RationaleResultSource[keyof typeof RationaleResultSource];


export const RationaleResultSource = {
  'deterministic-fallback': 'deterministic-fallback',
  ai: 'ai',
} as const;

export interface RationaleResult {
  source: RationaleResultSource;
  markdown: string;
}