# lib/api-zod/src/generated/types/rationaleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MetroFlow AI offline traffic engineering advisory API
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationMetrics } from './simulationMetrics';

export interface RationaleInput {
  intersectionId: string;
  peakWindow: string;
  vcRatio: number;
  cycleBefore: number;
  cycleAfter: number;
  phaseBefore: number;
  phaseAfter: number;
  delayBefore: number;
  delayAfter: number;
  safetyStatus: string;
  demand?: number;
  capacity?: number;
  queue?: number;
  currentSignal?: string;
  recommendedSignal?: string;
  simulationBefore?: SimulationMetrics;
  simulationAfter?: SimulationMetrics;
  assumptions?: string[];
}