# lib/api-zod/src/generated/types/simulationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MetroFlow AI offline traffic engineering advisory API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationInput {
  /**
     * @minimum 0.7
     * @maximum 1.3
     */
  demandMultiplier: number;
  baselineDelay?: number;
  baselineQueue?: number;
  baselineThroughput?: number;
  recommendedDelay?: number;
  recommendedQueue?: number;
  recommendedThroughput?: number;
}